requirements.txt
django-heroku
gunicorn
pydub
SpeechRecognition
keras
tensorflow
pickle
unidecode